Extracted the render of people in directory into its own functional component relates to #38
